+++ client/src/Pages/HomePage.jsx
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import Button from 'material-ui/Button';
 import SearchBar from '../components/home/SearchBar.jsx'
-import LocationMenu from '../components/home/DropDown.jsx'
+import LocationMenu from '../components/home/LocationMenu.jsx'
 
 const Home = (props) => (
   <div>
